Typecast the return value of sizeof from "unsigned long int" to "int".  This was giving a type error when printed on some other platforms (using g++).

+++ precolor_extend.cpp
@@ -516,7 +516,7 @@ int main(int argc, char *argv[])
         
         if (G->n > 8*sizeof(BIT_MASK))
         {
-            printf("The number n=%d of vertices in the graph is more than the number %d of bits in a word.",G->n,8*sizeof(BIT_MASK));
+            printf("The number n=%d of vertices in the graph is more than the number %d of bits in a word.",G->n,(int)(8*sizeof(BIT_MASK)));
             exit(9);
         }
